Added DI container and SQLite database

+++ KR1/KR1/Database/IFinanceDatabase.cs
@@ -6,23 +6,23 @@ namespace KR1.Database
     {
         void AddAccount(string name, decimal balance);
         void AddCategory(string name, bool isIncome);
-        void AddOperation(bool isIncome, int bankAccountId, decimal amount, DateOnly date, int categoryId, string description = "");
+        void AddOperation(bool isIncome, Int64 bankAccountId, decimal amount, DateOnly date, Int64 categoryId, string description = "");
 
         IEnumerable<BankAccount> GetAccounts();
         IEnumerable<Category> GetCategories();
         IEnumerable<Operation> GetOperations();
 
-        void RemoveAccount(int accountId);
-        void RemoveCategory(int categoryId);
-        void RemoveOperation(int operationId);
+        void RemoveAccount(Int64 accountId);
+        void RemoveCategory(Int64 categoryId);
+        void RemoveOperation(Int64 operationId);
 
-        void UpdateAccountName(int accountId, string newName);
-        void UpdateCategoryName(int categoryId, string newName);
-        void UpdateCategoryIsIncome(int categoryId, bool isIncome);
-        void UpdateOperationBankAccountId(int operationId, int newBankAccountId);
-        void UpdateOperationCategoryId(int operationId, int newCategoryId);
-        void UpdateOperationAmount(int operationId, decimal newAmount);
-        void UpdateOperationDate(int operationId, DateOnly newDate);
-        void UpdateOperationDescription(int operationId, string newDescription);
+        void UpdateAccountName(Int64 accountId, string newName);
+        void UpdateCategoryName(Int64 categoryId, string newName);
+        void UpdateCategoryIsIncome(Int64 categoryId, bool isIncome);
+        void UpdateOperationBankAccountId(Int64 operationId, Int64 newBankAccountId);
+        void UpdateOperationCategoryId(Int64 operationId, Int64 newCategoryId);
+        void UpdateOperationAmount(Int64 operationId, decimal newAmount);
+        void UpdateOperationDate(Int64 operationId, DateOnly newDate);
+        void UpdateOperationDescription(Int64 operationId, string newDescription);
     }
 }
