+++ KR1/KR1/Domain objects/Operation.cs
@@ -1,13 +1,13 @@
 ﻿namespace KR1.DomainObjects
 {
-    internal class Operation(int id, bool isIncome, int bankAccountId, decimal amount, DateOnly date, int categoryId, string description = "")
+    internal class Operation(Int64 id, bool isIncome, Int64 bankAccountId, decimal amount, DateOnly date, Int64 categoryId, string description = "")
     {
-        public int Id { get; private set; } = id;
+        public Int64 Id { get; private set; } = id;
         public bool IsIncome { get; set; } = isIncome;
-        public int BankAccountId { get; set; } = bankAccountId;
+        public Int64 BankAccountId { get; set; } = bankAccountId;
         public decimal Amount { get; set; } = amount;
         public DateOnly Date { get; set; } = date;
-        public int CategoryId { get; set; } = categoryId;
+        public Int64 CategoryId { get; set; } = categoryId;
         public string Description { get; set; } = description;
     }
 }
